# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Ecosystem API — CNC Tool Life",
    "description": "Backend ของระบบ CNC Tool Life AI: พยากรณ์อายุดอกกัด (RUL, time series) → ถอดดอก → วัดรอยสึก VB จากภาพใบมีด → ผู้ตรวจยืนยัน → ใบสั่งงาน → retrain\n\nทุก endpoint อยู่ใต้ `/api/v1` · ใช้ปุ่ม **Authorize** ใส่ Bearer token สำหรับ endpoint ที่ต้องล็อกอิน",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tool-vision/stations": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Stations",
        "operationId": "stations_api_v1_tool_vision_stations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/stations/{machine}/capture": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Capture",
        "description": "สำรองกรณีถ่ายภาพอัตโนมัติไม่สำเร็จ (เช่น แบบจำลองภาพยังไม่พร้อมตอนถอดดอก)",
        "operationId": "capture_api_v1_tool_vision_stations__machine__capture_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "machine",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Machine"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/inspections": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Inspections",
        "operationId": "inspections_api_v1_tool_vision_inspections_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "machine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/inspections/{ins_id}": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Inspection",
        "operationId": "inspection_api_v1_tool_vision_inspections__ins_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ins_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ins Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/inspections/{ins_id}/blades/{blade}/image": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Blade Image",
        "operationId": "blade_image_api_v1_tool_vision_inspections__ins_id__blades__blade__image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ins_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ins Id"
            }
          },
          {
            "name": "blade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Blade"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/inspections/{ins_id}/blades/{blade}/measure": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Measure",
        "operationId": "measure_api_v1_tool_vision_inspections__ins_id__blades__blade__measure_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ins_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ins Id"
            }
          },
          {
            "name": "blade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Blade"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/inspections/{ins_id}/review": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_tool_vision_inspections__ins_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ins_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ins Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/replacements": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Replacements",
        "operationId": "replacements_api_v1_tool_vision_replacements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/replacements/export/csv": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Replacements Csv",
        "operationId": "replacements_csv_api_v1_tool_vision_replacements_export_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/replacements/{ins_id}/done": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Replaced",
        "description": "ช่างเปลี่ยน/ลับดอกตามใบสั่งงานแล้ว (ปิดทั้งดอก)",
        "operationId": "replaced_api_v1_tool_vision_replacements__ins_id__done_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ins_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ins Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tool-vision/stats": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Stats",
        "operationId": "stats_api_v1_tool_vision_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/model": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Model",
        "operationId": "model_api_v1_tool_vision_model_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/model/versions": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Model Versions",
        "operationId": "model_versions_api_v1_tool_vision_model_versions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/model/activate": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Model Activate",
        "description": "ใช้เวอร์ชันที่เลือกเป็นตัวหลัก (ย้อนเวอร์ชันได้) — ตรวจ sha256 + self-test ก่อนสลับ (admin)",
        "operationId": "model_activate_api_v1_tool_vision_model_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/model/reload": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Model Reload",
        "operationId": "model_reload_api_v1_tool_vision_model_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/training/pool": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Training Pool",
        "operationId": "training_pool_api_v1_tool_vision_training_pool_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/training/jobs": {
      "get": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Training Jobs",
        "operationId": "training_jobs_api_v1_tool_vision_training_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/training/start": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Training Start",
        "operationId": "training_start_api_v1_tool_vision_training_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tool-vision/training/jobs/{job_id}/{action}": {
      "post": {
        "tags": [
          "Tool Vision (blade inspection)"
        ],
        "summary": "Training Decide",
        "operationId": "training_decide_api_v1_tool_vision_training_jobs__job_id___action__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivateBody": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "version"
        ],
        "title": "ActivateBody"
      },
...
      "BladeDecision": {
        "properties": {
          "blade": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Blade"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "AI = ยอมรับค่าที่ AI วัด · BENCH = ค่าจาก optical bench · MANUAL = ค่าที่วัดเอง",
            "default": "AI"
          },
          "vb_um": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vb Um",
            "description": "VB (µm) — ใช้เมื่อ source = MANUAL"
          }
        },
        "type": "object",
        "required": [
          "blade"
        ],
        "title": "BladeDecision"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewBody": {
        "properties": {
          "blades": {
            "items": {
              "$ref": "#/components/schemas/BladeDecision"
            },
            "type": "array",
            "title": "Blades"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "blades"
        ],
        "title": "ReviewBody"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}